Updates: changes for version 0.2.0
