+++ README.md
@@ -1,1 +1,12 @@
 # Todo CLI Application
+
+## Preview
+
+Adding a todo
+![ADD_SCREENSHOT](./img/todo-rust-cli-app[ADD_SCREENSHOT].PNG)
+
+Displaying the help menu
+![HELP_MENU_SCREENSHOT](./img/todo-rust-cli-app[HELP_MENU_SCREENSHOT].PNG)
+
+Showing all todos
+![SHOW_SCREENSHOT](./img/todo-rust-cli-app[SHOW_SCREENSHOT].PNG)
